pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "folium>=0.19.4",
    "geopy>=2.4.1",
    "spacy>=3.8.4",
    "streamlit-folium>=0.24.0",
    "streamlit>=1.42.0",
    "youtube-transcript-api>=0.6.3",
    "twilio>=9.4.5",
    "trafilatura>=2.0.0",
    "google-generativeai>=0.8.4",
    "openai>=1.63.0",

   
]
[theme]
base="light"  # veya "dark"

[browser]
serverAddress = "localhost"

[server]
enableStaticServing = false

[client]
showErrorDetails = false
